Demoqa_automation/requirements.txt
paramiko
selenium
pytest
pyyaml
paramiko
selenium
pytest
pyyaml
pytest-html


